+++ CHANGELOG.md
@@ -2,5 +2,8 @@
 - Fix ESP8266 stale crash
 - Fix unit tests
 - Add screen status handler on start
+- Add Github action to check version
+- Add version.json file
+- Add device value in platform
 
 - Add auto OTA update check on start